requirements.txt
requests
tabula-py
boto3